[Feature] Refactor Features section: streamline component structure and enhance Hero section
