--- src/apps/web/src/widgets/home/Features/types/FeatureColor.ts
@@ -1,4 +1,0 @@
-/**
- * Accent colors for feature cards
- */
-export type FeatureColor = "blue" | "yellow";
