--- src/apps/web/src/widgets/home/Hero/components/HeroActions/index.ts
@@ -1,2 +1,0 @@
-export { default, HeroActions } from "./HeroActions";
-
